+++ src/Oro/Bundle/GoogleTagManagerBundle/Resources/public/js/app/components/analytics4/image-slider-gtm-analytics4-component.js
@@ -1,46 +1,46 @@
 define(function(require) {
-    const BaseComponent = require('orogoogletagmanager/js/app/components/base-embedded-list-gtm-component');
+    const BaseEmbeddedListGtmComponent = require('orogoogletagmanager/js/app/components/base-embedded-list-gtm-component');
     const mediator = require('oroui/js/mediator');
     const _ = require('underscore');
 
     /**
      * Listens to oro:embedded-list:* events and invokes promo "select_promotion, "view_promotion" GTM GA4 events
      */
-    const HomePageSliderGtmAnalytics4Component = BaseComponent.extend({
+    const ImageSliderGtmAnalytics4Component = BaseEmbeddedListGtmComponent.extend({
         /**
          * @property {Object}
          */
-        options: _.extend({}, BaseComponent.prototype.options, {
-            nameSelector: '[data-role="slide-content"] h2'
+        options: _.extend({}, BaseEmbeddedListGtmComponent.prototype.options, {
+            nameSelector: '[data-role="slide-item-link"] img'
         }),
 
         /**
          * @inheritdoc
          */
-        constructor: function HomePageSliderGtmAnalytics4Component(options) {
-            HomePageSliderGtmAnalytics4Component.__super__.constructor.call(this, options);
+        constructor: function ImageSliderGtmAnalytics4Component(options) {
+            ImageSliderGtmAnalytics4Component.__super__.constructor.call(this, options);
         },
 
         /**
          * @inheritdoc
          */
-        _invokeEventView: function(viewData) {
+        _invokeEventView(viewData) {
             mediator.trigger('gtm:event:analytics4:view_promotion', viewData);
         },
 
         /**
          * @inheritdoc
          */
-        _getModel: function($item) {
+        _getModel($item) {
             return {
-                name: $item.find(this.options.nameSelector).text().trim()
+                name: $item.find(this.options.nameSelector).attr('alt')
             };
         },
 
         /**
          * @inheritdoc
          */
-        _getViewData: function(model, index) {
+        _getViewData(model, index) {
             return {
                 creative_name: this._getBlockName(),
                 item_name: model['name'],
@@ -51,17 +51,17 @@ define(function(require) {
         /**
          * @inheritdoc
          */
-        _invokeEventClick: function(clicksData, destinationUrl) {
+        _invokeEventClick(clicksData, destinationUrl) {
             mediator.trigger('gtm:event:analytics4:select_promotion', clicksData, destinationUrl);
         },
 
         /**
          * @inheritdoc
          */
-        _getClickData: function(model, index) {
+        _getClickData(model, index) {
             return this._getViewData(model, index);
         }
     });
 
-    return HomePageSliderGtmAnalytics4Component;
+    return ImageSliderGtmAnalytics4Component;
 });