+++ src/Oro/Bundle/GoogleTagManagerBundle/Resources/public/js/app/components/data-layer-manager-component.js
@@ -1,107 +1,78 @@
-define(function(require) {
-    'use strict';
+import BaseComponent from 'oroui/js/app/components/base/component';
+import mediator from 'oroui/js/mediator';
 
-    const BaseComponent = require('oroui/js/app/components/base/component');
-    const mediator = require('oroui/js/mediator');
-    const $ = require('jquery');
-    const _ = require('underscore');
+mediator.setHandler('gtm:data-layer-manager:isReady', () => false);
 
-    const DataLayerManagerComponent = BaseComponent.extend({
-        /**
-         * @property {Object}
-         */
-        options: _.extend({}, BaseComponent.prototype.options, {
-            dataLayerName: ''
-        }),
+const DataLayerManagerComponent = BaseComponent.extend({
+    /**
+     * @property {Object}
+     */
+    options: Object.assign({}, BaseComponent.prototype.options, {
+        dataLayerName: ''
+    }),
 
-        /**
-         * @property {Array}
-         */
-        _dataLayer: [],
+    /**
+     * @property {Array}
+     */
+    _dataLayer: null,
 
-        /**
-         * @property {Object}
-         */
-        listen: {
-            'gtm:event:push mediator': '_onPush'
-        },
+    /**
+     * @property {Object}
+     */
+    listen: {
+        'gtm:event:push mediator': '_onPush'
+    },
 
-        /**
-         * @inheritdoc
-         */
-        constructor: function DataLayerManagerComponent(options) {
-            DataLayerManagerComponent.__super__.constructor.call(this, options);
-        },
+    /**
+     * @inheritdoc
+     */
+    constructor: function DataLayerManagerComponent(options) {
+        DataLayerManagerComponent.__super__.constructor.call(this, options);
+    },
 
-        /**
-         * @param {Object} options
-         */
-        initialize: function(options) {
-            DataLayerManagerComponent.__super__.initialize.call(this, options);
+    /**
+     * @param {Object} options
+     */
+    initialize(options) {
+        DataLayerManagerComponent.__super__.initialize.call(this, options);
+        this.options = Object.assign({}, this.options, options || {});
+        this._dataLayer = window[this.options.dataLayerName] || [];
 
-            this.options = _.defaults(options || {}, this.options);
-        },
+        this._deferredInit();
+        const onGTMLoaded = () => {
+            mediator.setHandler('gtm:data-layer-manager:isReady', () => true);
+            this._resolveDeferredInit();
+        };
+        const gtmScript = document.querySelector('script[data-gtm-integration]');
 
-        /**
-         * @inheritdoc
-         */
-        delegateListeners: function() {
-            DataLayerManagerComponent.__super__.delegateListeners.call(this);
-
-            const gtmLoaded = $.Deferred();
-            const dataLayer = window[this.options.dataLayerName];
-            if (dataLayer instanceof Array && dataLayer.push !== Array.prototype.push) {
-                // Google Tag Manager has been already loaded.
-                gtmLoaded.resolve();
-            } else {
-                window.addEventListener('gtm:loaded', function() {
-                    gtmLoaded.resolve();
-                }, {once: true});
-            }
-
-            const pageLoaded = $.Deferred();
-            mediator.once('page:afterChange', function() {
-                pageLoaded.resolve();
-            });
-
-            // Wait until GTM is loaded and page is fully ready.
-            $.when([gtmLoaded, pageLoaded]).done((function() {
-                // Copy dataLayer contents if we already have something queued to push to data layer.
-                const dataLayerOld = this._dataLayer;
-
-                this._dataLayer = window[this.options.dataLayerName];
-
-                if (dataLayerOld.length) {
-                    // Push queued content to data layer.
-                    _.each(dataLayerOld, (function(item) {
-                        this._onPush(item);
-                    }).bind(this));
-                }
-
-                mediator.trigger('gtm:data-layer-manager:ready');
-            }).bind(this));
-        },
-
-        /**
-         * @returns {Array}
-         */
-        getDataLayer: function() {
-            return this._dataLayer;
-        },
-
-        /**
-         * @param {Object} data
-         * @param {Boolean} [clear] Clear ecommerce object before pushing data. True by default.
-         * @private
-         */
-        _onPush: function(data, clear = true) {
-            if (clear) {
-                this.getDataLayer().push({ecommerce: null});
-            }
-
-            this.getDataLayer().push(data);
+        if (
+            // there's no gtm.js script, in case the integration is stubbed for a test
+            !gtmScript ||
+            // gtm.js already loaded
+            gtmScript.loadDone ||
+            // gtm.js is failed to load, resolve deferredInit to unblock UI (there's already system error in console)
+            gtmScript.loadError
+        ) {
+            onGTMLoaded();
+        } else {
+            window.addEventListener('gtm:loaded', onGTMLoaded, {once: true});
+            // gtm.js is failed to load, resolve deferredInit to unblock UI (there's already system error in console)
+            window.addEventListener('gtm:error', onGTMLoaded, {once: true});
         }
-    });
+    },
 
-    return DataLayerManagerComponent;
+    /**
+     * @param {Object} data
+     * @param {boolean=} clear Clear ecommerce object before pushing data. True by default.
+     * @private
+     */
+    _onPush(data, clear = true) {
+        if (clear) {
+            this._dataLayer.push({ecommerce: null});
+        }
+
+        this._dataLayer.push(data);
+    }
 });
+
+export default DataLayerManagerComponent;